@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.1.1+1
 
 environment:
-  sdk: ">=2.18.6 <3.0.0"
+  sdk: '^3.0.0'
 
 dependencies:
   flutter:
@@ -31,9 +31,9 @@ dependencies:
   http: ^0.13.3
   localstorage: ^4.0.0+1
   flutter_multiselect: ^1.0.0
-  flutter_svg: ^1.1.1+1
+  flutter_svg: ^2.0.5
   url_launcher: ^6.0.9
-  flutter_slidable: ^2.0.0
+  flutter_slidable: ^3.0.0
   fluttertoast: ^8.0.8
   bmprogresshud: 1.0.1
   email_validator: ^2.0.1
